backend/requirements.txt
fastapi==0.115.14
google-cloud-speech==2.21.0
numpy==1.22.0
protobuf==6.31.1
psycopg2_binary==2.9.10
python-dotenv==1.1.1
sentence_transformers==5.0.0
uvicorn==0.35.0